Refactor location of trace helpers.
