+++ src/v2/trace.ts
@@ -1,36 +1,7 @@
-import { AsyncLocalStorage } from 'async_hooks';
 import * as express from 'express';
 
+import { TraceParent, getTraceParent, traceContext } from '../common/trace';
 import { CloudEvent } from './core';
-
-const traceContext = new AsyncLocalStorage<TraceParent>();
-
-const TRACE_PARENT_REGEX = new RegExp(
-  '^(?<version>[\\da-f]{2})-' +
-    '(?<traceId>[\\da-f]{32})-' +
-    '(?<parentId>[\\da-f]{16})-' +
-    '(?<flag>[\\da-f]{2})$'
-);
-
-interface TraceParent {
-  version: string;
-  traceId: string;
-  parentId: string;
-  flag: string;
-}
-
-function getTraceParent(carrier: any): TraceParent | undefined {
-  const traceParent = carrier?.['traceparent'];
-  if (!traceParent) {
-    return;
-  }
-  const matches = TRACE_PARENT_REGEX.exec(traceParent);
-  if (!matches || !matches.groups) {
-    return;
-  }
-  const { version, traceId, parentId, flag } = matches.groups;
-  return { version, traceId, parentId, flag };
-}
 
 type HttpsFunction = (
   req: express.Request,
